Moved the new api folder to the server and removed unused views.
