+++ src/server/app_api/controllers/mainControllers.js
@@ -1,0 +1,3 @@
+module.exports.index = function(req, res) {
+    res.json({"message" : "API online"});
+}
