+++ src/server/app_api/routes/index.js
@@ -3,14 +3,19 @@ const router = express.Router();
 const ctrlLocations = require('../controllers/locations');
 const login = require('../controllers/login');
 const register = require('../controllers/register');
+const mainController = require('../controllers/mainControllers');
+
+router
+.route('/')
+.get(mainController.index)
 
 // locations
 router
-.route('/locations') 
+.route('/locations')
+.get(ctrlLocations.locationsCreate)
+//.get(ctrlLocations.locationsListByDistance)
+//.post(ctrlLocations.locationsCreate);
 
-.get(ctrlLocations.locationsCreate) 
-//.get(ctrlLocations.locationsListByDistance) 
-//.post(ctrlLocations.locationsCreate); 
 router
 .route('/login')
 .post(login.loginCreate)
@@ -19,11 +24,11 @@ router
 .route('/register')
 .post(register.registerCreate)
 
-/* router 
-.route('/locations/:locationid') 
+/* router
+.route('/locations/:locationid')
 
-.get(ctrlLocations.locationsReadOne) 
-.put(ctrlLocations.locationsUpdateOne) 
+.get(ctrlLocations.locationsReadOne)
+.put(ctrlLocations.locationsUpdateOne)
 .delete(ctrlLocations.locationsDeleteOne); */
 
 module.exports = router;
