+++ src/server/app_api/views/layout.pug
@@ -1,0 +1,6 @@
+doctype html
+html
+    head
+        title= title
+    body
+        block content
