--- src/server/app_server/views/includes/scheduleTable.pug
@@ -1,39 +1,0 @@
-table(class="table")
-    thead
-        th Exercise
-        th Description
-        th Set
-        th Reps/time
-        th
-            form(class = "form-group", action = "/deleteSchedule", method = "POST")
-                input(type="hidden", name="scheduleId", value="" + schedule.id)
-                button(class = "btn btn-default float-right" type="submit")
-                    i(class="fas fa-trash-alt")
-    tbody
-        each exercise in schedule.exercises
-            tr
-                td=exercise.exerciseName
-                td=exercise.description
-                td=exercise.set
-                td=exercise.reps
-                td
-                    form(class = "form-group", action = "/deleteScheduleExercise", method = "POST")
-                        input(type="hidden", name="scheduleId", value="" + schedule.id)
-                        input(type="hidden", name="exerciseId", value="" + exercise.id)
-                        button(class = "btn btn-default float-right")
-                            i(class="fas fa-times")
-form(class="form-group", action="/postScheduleItem", method="POST")
-    input(type="hidden", name="scheduleId", value="" + schedule.id)
-    table(class="table table-borderless")
-        tr
-            td
-                input(class="form-control", name="exerciseName", type="text", placeholder="Exercise", required=true)
-            td
-                input(class="form-control", name="description", type="text", placeholder="Description", required=true)
-            td
-                input(class="form-control", name="set", type="number", value=0)
-            td
-                input(class="form-control", name="reps", type="number", value=0)
-            td
-                button(class="btn btn-default" type="submit")
-                    i(class="fas fa-plus")
