--- src/server/app_server/views/index.pug
@@ -1,11 +1,0 @@
-extends layout
-
-block content
-    h1 Welcome to #{title}
-    div(class="row")
-        each schedule in schedules
-            div(class="col-md-6")
-                    include includes/scheduleTable
-        div(class="col-md-6")
-            form(class="form-group", action="/postSchedule", method="POST")
-                button(class="add-schedule-button", type="submit") Create a new schedule
