--- src/server/public/stylesheets/style.css
@@ -1,20 +1,0 @@
-body {
-    padding: 50px;
-    font: 14px "Lucida Grande", Helvetica, Arial, sans-serif;
-}
-
-a {
-    color: #00B7FF;
-}
-
-.add-schedule-button {
-    width: 100%;
-    border: dashed black 2px;
-    border-radius: 10px;
-    height: 50px;
-}
-
-.add-schedule-button:hover{
-    background-color: lightgray;
-    cursor: pointer;
-}
